fix: go build target with makefile
With `go build` we can't specify more than one binaries to build. Go
will not know what to do with them. It feels logical to assume it will
build a binary with the same name as the package defined, but
1. You can specify `-o` flag and it can only be one.
2. You can have different modules with the same name like `cmd/mybin`
   and `extra/mybin`.

To prevent polluting the root of the repo, I moved all binary under
the `bin/` directory.