incorporate review
Additionally I found a strange behaviour of the yaml path library.
It cannot handle mixed yaml json. The resulting output when executing
a modification is a completely corrupted yaml.

Now, the susbtitution package wraps this lib by feeding it with
pure json or yaml, only, dependening on the original input format.