fix: address gosec issues (#123)
* fix(dx): address gosec issues

Fixed issues:

* G303: File creation in shared tmp directory without using ioutil.Tempfile (gosec)
* G404: Use of weak random number generator (math/rand instead of crypto/rand) (gosec)
* G110: Potential DoS vulnerability via decompression bomb (gosec)
* G601: Implicit memory aliasing in for loop. (gosec)

About G303
----------

Related code was removed. It was and left there most likely for debugging.

About G404
----------

Right now I don't see any usage of a publicly accessible `RandomString`
function:

```
❯ rg 'RandomString'
pkg/utils/utils.go
116:// RandomString creates a new random string with the given length.
117:func RandomString(n int) string {
```

Even if it's not intended to be used in a cryptographic context, it's an
exported function and uses a random value generator, therefore it should use a
secure generator.

About G110
----------

```go
tr := tar.NewReader(in)
if _, err := io.Copy(file, tr); err != nil { }
```

That can be a serious security issue because `Copy` reads until `EOF` or
`error`, but it is possible to create a decompression bomb that can cause a
denial of service as it will never stop reading/writing.

The resolution would be to use `CopyN` instead, as a result we have an limit on
the maximum size of an archive.

This may looks like "nah it's fine, i can press ctrl+c", but on a controller it
can cause a lot of issues, or an automated process tries to deploy something on
a fenced network, somehow the malicious tar referenced file passed security
checks, so it will bring down the whole bare metal machine on a
nuclear-submarine.

We don't know what size limit we could set, the tar archive can be an
image layer and that can even reach the gigabyte range. For now, we
acknowledge the risk.

We checked other softwares and tried to figure out how they manage this,
but it's handled the same way.

Related answer on SO: https://stackoverflow.com/a/67330806

About G601
----------

All the reference changes may look useless, but I'll try to demonstrate why I
made these changes and why gosec complains about them.

Passing down a variable as a reference can lead to unexpected failures as the
referred entity can be changed after it was passed to a function. If the function
for example is on a struct and the struct stores that reference, it may have
difference value later on when it tries to use it. As the code gets more
complex, it can cause issues that can be debugged for hours or days to find
why given value is a nil reference "sometimes".

In a for loop variables are destroyed at the end of the loop, but if something
refers to them, the GC will not removed them and they will live "forever" in
the memory leading to memory leaks, for example here:

```go
for _, tag := range tags {
  n := *a
  n.Tag = tag
  r.addArtefactInfo(&n)
}
```
One can argue passing whole structs between functions can be memory expensive,
but potential memory leaks and race conditions are more harmful, and if a
function gets a reference as an argument, they can change it which can lead to
malfunctions through side-effects.

There are cases where we want to use references in a for loop, but that should
be an edge case and not a standard.

References
----------

* https://github.com/open-component-model/ocm/issues/86
* https://github.com/securego/gosec
* https://golangci-lint.run/usage/linters/

Closes #86

* use rand.Intn as fallback for crypro rand to ensure valid output

harmonize component descriptor validation and conversion.

Co-authored-by: Uwe Krueger <uwe.krueger@sap.com>