fix panic in tests
decouple the processing code from ocm context by changing logging
directly to the logging context, instead of requireing
an ocm context, whose other functionality is not used here.

Provide a logging context defaulting to solve the panic
occurring during the initialization, when output objects
are intantiated to figure out its interface to add the
appropriate command line options. This lead to the usage
of a nil context, which caused the panic.

fix unintended rename of symbol in hack/format.sh