fix adding git sources without configured credentials (#142)
if (re)sources are added to a component archive the access method is
created for validation. This must be possible, even if the technical
access requires crednetials, which are not yet provided.

Therefore operations requiring credentials (for private repos)
for the github access method have been moved to the methods
really requesting access to the content.