support for deferred error propagation
A typical problem is to handle errors provided by deferred functions,
for example Close().

Another scenario is the need to wrap all errors provided in a function
with a common context description.

The errors package now provides a function PropagateError, which
can be used to solve both problems.

If takes an optional error context, which will be used to wrap
a potential error returned by the calling function.
Additionally it is possible to add a function providing
an error. This error, if present is composed with an error
provided by the function return when called as deferred function.

Example:

func() (efferr error) {
  ...
  defer errors.PropagateErrorf(&efferr, stream.Close, "error context")
  ...
  return err
}

This error providing function may be the Finalize() method of a
utils.Finalizer.

This is now directly supported by the Finalizer.

Example:

func() (efferr error) {
  var finalize utils.Finalizer
  ...
  defer finalize.FinalizeWithErrorPropagationf(&efferr, "error context")
  ...
  finalize.Close(stream)
  ...
  return err
}