consistent view handling for access methods
eliminate the hybrid handling for access methods:
- as provided by the access type: non-dup
- explicit method view interface with dup and
  wrapper introducing the dup.

Now, the cosumer view is just the AccessMethos interface
providing ref counting and Dup.

The old interface is just the implementation interface
AccessMethodImpl, now. It must be wrapped ba an access method provider
using a cpi function into a regular AccessMethod object before
returning it to the consumer.

ATTENTION: for regular consumers of the intreface, this is transparent,
but type casts to the implementation type is not possible anymore.
If there is internal coding urgently requiring access to the
implementation, there is a cpi function GetAccessMethodImplementation.