Use cache implementation in session
The session implementation implemented its own repository cache which lead to code duplication and the reoccurrence of an error that was already fixed for the other implementation. The session has to take over responsibility for open repositories because the objects opened during a session have to be closed in the reversed order they were opened (e.g. components opened based on a previously opened repository have to be closed first)