Remove finalization of repos from repo cache
Repository Cache is a utility object that can be used by several higher level objects. Since the closing of the repository objects depends on the usage context (e.g. if components have been looked up in this repository, these components have to be closed before the repository can be closed), it is the responsibility of the higher level objects to close the repositories correctly. To be able to determine, whether a repo has to be added to a finalizer or whether it has already been added, the cache also returns a bool indicating whether the repo was already cached (true) or not (false).