Vault tests (#749)
## Description
Implement integration tests with vault.

It also fixes a bug where getting the credentials for a consumer
provider lead to a deadlock or potentially an infinite recursion since
we attempt to retrieve credentials to access the credential provider
from the credential provider itself.

Introduces an additional make target called `make test-all`. This runs 
all test including the ones tagged with `go:build integration`.