(redo)Only re-encode subst value if it smells like json and target doc is yaml  (#796)
* Improve handling of substitution values wrt json or yaml encoding.

If target is json then of course value  must be json.   Not much to talk about there.

However if target is yaml and value is json then we would like to convert the json to yaml first so we don't have a target that looks like a mix of yaml and json.

That said, if value and target are both yaml then we would like to preserve whatever is in the value.   e.g. anchors and refs, string styles etc.

* In SubstituteByData, only re-encode value to yaml if the value bytes look like json and if the target file is not json.

* Only perform json sniff if target isn't json

* Fix small typo in sniffJson regex

---------

Co-authored-by: Hilmar Falkenberg <hilmar.falkenberg@sap.com>