introduce optional repository interface for resolvers (#867)
#### What this PR does / why we need it:

CLI command `ocm list componentversions` dumps if only a component is
specified together with a resolver (either as part of the config or as
option).

This PR fixes this dump.

The resolvers are only able to resolve component versions. Therefore,
after the fix the provided list does not show any version for the given
component. The expected behaviour would surely be to get the list of
versions resolvable by the given resolvers.

Not all resolvers are able to provide repositories offering a version
list method. Therefore an optional interface is introduced for
resolvers, which provide potential repositories for a given component.
This interface offers the possibility to return repositories or abstract
component providers which can be used to get versions or version lists.

This is then checked by the list handling to gather versions provided by
all found repositories.

#### Which issue(s) this PR fixes:
Fixes: #864