Resource resolution on CD set (#865)
#### What this PR does / why we need it:

Introduction resolution of relative resource references of the level of
component descriptors.

So far, this the resolution was only possible on the level of the
complete component version access abstraction.
This PR adds such a functionality on the level of plain component
descriptors.
Therefore, it provides a `ComponentVersionResolver` abstraction for
component descriptors, with two
implementations: a set of component descriptors and a compound resolver
similar to the one for component version accesses.

To align the element names in package `compdesc`, the
`ComponentReference`ans been renamed to `Reference` (like `Resource` and
`Source`). The old type has been deprecated.

This functionality is required for the OCM controllers working an deep
resources for a component version.

To verify the digest of the retrieved content of a resource, a new
function
`signing.VerifyResourceDigest(cv, index, content)`
is provided. it uses the digesting type described by the component
version to recalculate the digest of the retrieved resource and verify
it against the digest described by the component version.