Custom Unmarshaler for Consumer Identity (#927)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
This PR adds a custom unmarshaler to the consumer identity type.

Adding this custom unmarshaler to the consumer identity allows a yaml
specification containing a data type other string, e.g. a hostpath spec
with a port. Previously, it would error if the user specified `port:
5000` and instead, the user had to specify `port: "5000"`.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->