fix closure + digest mode + ctf index (#973)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

- Calling `ocm hash components -U` provides hashes as top-level state,
and not on the level of component references.
The reason is, that the digest mode is not initialized in the signing
options.

- The index of the OCI ctf does not handle tag updates correctly. It
does not remove the old digest to tag mapping.

- The Recursion modes of the CLI does not work correctly for closable
elements. The Exploder had no access to the session hasn't added the
injected new elements to the session. Therefore the repository could not
be closed, which left the index state unchanged on the filesystem. The
error from the deferred close was ignores .

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->