chore: release fallout corrections (#1023)
runs make generate because it was only run on the release branch. adds
check as mandatory dep of release to make sure the tag-check inside is
not rgiacing against release. adds necessary build steps for tools
needed in make generate.

<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->