chore: release branches as releases/vX.Y
This reworks the release branch action to correctly branch off in patterns that do not include Z / Patch versions:

```
releases/v0.17.0 # before
releases/v0.17 # after
```

To achieve this it allows to print only the major/minor combination for the release_generate.go generator.

Additionally, it adds an automatic bump to main after the Branch Cutoff was completed into main.