fix implicit version handling of an artifact identity (#1026)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
The defaulting of the extraIdentity did only work, if there was an
identity map already set.

So far the version was implicutly added to the extraIdentity of a CV if
the rest is not
unique. In the future the extraIdentity should be explicitly set to be
unique. Therefore,
the version was now implicitly added to the extraIdentity. This has two
problems:
- it was only done, if there was already an identity map set,
- it was always done, when serializing a CD

To be comparable with older signed component versions, such defaulting
may only be done
if the content of a component version is changed, otherwise the
signature would be brocken.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->