chore: fixup release action versioning and notes process (#1124)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Makes sure that release notes are correctly drawn in and that we always
have the version set correctly in our CTFs.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

This fixes up the release action in 2 important ways:

1. As part of the removal of the release notes we now no longer have a
file to use for release notes to give to goreleaser. This means we have
to temporarily create one
2. While discovering this I noticed that the Component Versions that are
built, are actually using the `VERSION` variable while
`EFFECTIVE_VERSION` is a combination of VERSION+COMMIT by default. I
didnt want to change this behavior originally, so now I started to use
VERSION for the Makefiles (which also means I had to export them as
settable). This now leads to correct CTF versions `v0.19.0-rc.1` instead
of `v0.19.0-dev` for example.

As such its a follow up from
https://github.com/open-component-model/ocm/pull/1120 and
https://github.com/open-component-model/ocm/pull/1077