chore: fixup release notes
the release notes that were fixed up require a small adjustment to be able to deal with properly multilined files.

This is because the notes are stored in bash variables which removes newlines by default. Now they are properly formatted in again and we also fixup the script so that the line count is correct