chore(backport): fixup release notes (#1148) (#1149)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

the release notes that were fixed up require a small adjustment to be
able to deal with properly multilined files.

This is because the notes are stored in bash variables which removes
newlines by default. Now they are properly formatted in again and we
also fixup the script so that the line count is correct

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`. -->

Makes sure that multilined release notes are interpreted correctly and
dont break out. This happened with release/v0.19 so that commit was
patched up by force push to test out the notes

<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->