feat: support splitting blobs when stored as OCI layer (#1140)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
There are OCI repositories with a layer size limitation. Because OCM
potentially maps any external artifact to a single blob stored as layer
on OCI registries, this could lead to problems. An obvious problematic
scenario is the transport of a multi-platform OCI image. Its blob format
is an archive containing all images and all layers of those images.

This PR introduces the possibility to specify blob limits for OCI
registries. The OCM-to-OCI mapping then
splits larger blobs into multiple layers.
The `localBlob` access method then uses a comma-separated list of blob
layer-blob digest to remember the sequence of layers.
The access then combines the layerblobs again to a single stream.
 
#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes: https://github.com/open-component-model/ocm-project/issues/12

A follow-up issue
https://github.com/open-component-model/ocm-project/issues/338 describes
the provisioning of appropriate defaults for common public registries.

---------

Co-authored-by: Jakob Möller <jakob.moeller@sap.com>