feat: add context to support cancel to tool ops (#1380)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This PR changes the internal tool operations for signing and transfer to
use a `context.Context`.
This context can be used to cancel the operations. The check is always
done before a new component version is
handled. This means, if cancelled, the actual component version handled
after a cancel will always be finished before
the operation is cancelled.

The consumer facing operations are added to provide versions accepting a
context.

#### Which issue(s) this PR fixes
Fixes #1346

---------

Co-authored-by: Jakob Möller <jakob.moeller@sap.com>