fix: make sure that tars generated by git access do not consider mod time
the git access method created blobs that would have differing digests and were not byte-compatible because of modtimes that were encoded. however, for ref'ed git repos, mod times are irrelevant.

additionally fixes that nested directories were not tar'ed correctly. This wasn't noticed because the test only contained flat files, this was now fixed