fix: git access tar methods (#1454)
The method to tar a filesystem that is currently used, also flattens the
directory structure to a simple list of files. As this is not what we
need, we will use another method.

Additionally, the tar-function includes the "last-modification" date of
the files. This leads to false results (e.g. digest calculation) as the
modification date for files is resetted when downloaded.

---------

Co-authored-by: Jakob Möller <jakob.moeller@sap.com>