fix: make sure that helm downloader does not overwrite chart with provenance data
in the helm downloader, there was a case when the chart data was overwritten with the provenance data.
if the target path passed to the downloader was a path such as `path.tgz` then this path was reused for writing the provenance file.

This was now fixed and there will be:

1. A download path pointing to the .tgz => chart.tgz
2. A provenance file (if it exists) => chart.prov

Note that before, due to buggy behavior the output provenance file was just called `chart` because the .prov file extension was not properly added. this has been fixed as well.

additionally if the file path was `tar.gz` instead of `tgz`, the file ending removal was incorrect. this has been fixed.