fix: make sure that helm downloader does not overwrite chart with provenance data (#1481)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

in the helm downloader, there was a case when the chart data was
overwritten with the provenance data. if the target path passed to the
downloader was a path such as `path.tgz` then this path was reused for
writing the provenance file.

This was now fixed and there will be:

1. A download path pointing to the .tgz => chart.tgz
2. A provenance file (if it exists) => chart.prov

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Note that before, due to buggy behavior the output provenance file was
just called `chart` because the .prov file extension was not properly
added. this has been fixed as well.

additionally if the file path was `tar.gz` instead of `tgz`, the file
ending removal was incorrect. this has been fixed.

This was also the last reason why the ocm download resource command
returned incorrect data in the tgz when downloading a resource with
accessType helm.

fix https://github.com/open-component-model/ocm/issues/1469

(cherry picked from commit c2e1bd2cfb20d7fe97415580013bad004858778e)