fix: double slash in `imageReference` (#1548)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This PR fixes the problem described below.

Steps to reproduce.
1. Prepare the following for the `ocm transfer cv` command:
    1.1. An OCM component having a resource of type `ociImage`
    1.2. An OCI registry to transfer the component to
2. Use the following parameters for the transfer operation:
`--copy-resources --uploader
'ocm/ociArtifacts={"ociRef":"<TARGET_OCI_REGISTRY>//"}'`

Note the problem:
1. The parameters instruct OCM CLI to use a named upload handler to copy
OCI artifacts to the root of the specified registry.
2. View OCM component descriptor of the copied component in the target
registry and note that the image reference to the OCI resource has `//`
between registry host and artifact namespace. The reference is invalid,
as there should be only one `/`.

---------

Signed-off-by: Ilya Khandamirov <ilya.khandamirov@sap.com>