fix: update JSON schema to allow null values for arrays in component descriptor V2 (#1645)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

this is a fix for a common issue in decoding descriptors that have
"null" values set in the JSON which was previously rejected. as this is
semantically equivalent to empty slices we want to allow this.

#### Which issue(s) this PR is related to
<!--
Usage: `Related to #<issue number>`, or `Related to (paste link of
issue)`.
-->

part of https://github.com/open-component-model/ocm-project/issues/680
as I noticed that null values are serialized normally but cannot be
deserialized correctly

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>