chore(deps): bump go version in flake (#1654)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Bump go version used in flake from 1.24 to 1.25.
#### Which issue(s) this PR is related to
<!--
Usage: `Related to #<issue number>`, or `Related to (paste link of
issue)`.
-->
Related to PR #1596

I noticed while creating this PR that there is already a PR tackling
this but it was created when go 1.25 was not yet in stable.
I hope you don't mind me opening a second PR

Signed-off-by: Tim <32556895+Avarei@users.noreply.github.com>
Co-authored-by: Jakob Möller <jakob.moeller@sap.com>